Refactored system action architecture.
Now an action is implemented as a lambda. Note that any closure in the lambda should be captured by value since they most likely have been destroyed by the time it's run.

This also makes the action namespace redundant and the different streams earlier implemented. Current implementation stores the actions (std::function) in a vector that can be resized if needed.